backend/core-service/requirements.txt
# Flask Core Service Dependencies

# Flask & Extensions
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5

# Database
sqlalchemy==2.0.23
psycopg[binary]>=3.1.0
alembic==1.13.0

# Authentication
pyjwt==2.8.0
werkzeug==3.0.1

# Kafka
kafka-python==2.0.2

# Redis
redis==5.0.1

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# AWS S3
boto3>=1.34.0

pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Development
black==23.12.0
flake8==6.1.0
